requirements.txt
python-telegram-bot
ccxt
pandas
pandas_ta
flask
gunicorn